Manually format 24 hour time in the time field
When calculating the default value for the time field, the
implementation now manually formats a string based on the hours,
minutes, and seconds attributes of the now() time object, instead of
using toLocaleTimeString as was previously the case.  This allows the
time to be formatted in a 24-hour format for all major browsers,
whereas toLocaleTimeString is only guaranteed to support whichever
format the current locale determines.