Manually format the default value for the date field
The default value for the date field is now calculated by manually
formatting a string wit the month, day, and year of the current date,
instead of using toLocaleDateSTring as was previously the case.  This
allows the implementation to make sure that the date is in the correct
format for the form without depending on the user's locale to be set
correctly.