Disable "other" controls when not in use
The controls that are used for entry of an "other" value in, for
example, the ICS position form fields are now disabled when the
combobox has not been set to "other".  This required a slight CSS
change to make sure that the background colors of disabled elements
are correct.  Finally, this also remedies the problem of duplicate
values being sent for form fields with "other" entry fields, as the
"other" field will be disabled and empty when the form is sent if the
combobox is not set to "other".