Fill in form fields from query string
This commit includes a mechanism to fill in form fields from the query
string using the Outpost names for the fields.  This should allow good
integration with outpost: selecting the form from the Outpost
menu (after suitable renaming) results in Outpost opening the form
with all of the correct values filled in.