Add array_for_each and use where needed
A function array_for_each which maps to the Array object prototype
forEach method is defined.  This allows forEach to be easily applied
to Array-like objects like DOM NodeLists.

All places where the direct call to the prototype method was being
used were replace with the shorter function name.  All calls were
formatted consistently with the array object and function prototype on
the same line with array_for_each and the function body indented
below.  Emacs properly indents this style.

The one place where indexed iteration was used in place of a forEach
was updated to use array_for_each as well.