Make form validation work when submitting to Opdirect
Previously, the Outpost/Opdirect submission used a second form for the
submit control so that the field of the second form (which contained
the serialized formdata) could be used to send the form to the
Opdirect server.  However, this meant that validation on the main form
did not happen before the form wast POSTed.  This commit moves the
button into the other form and gives it a handler that prevents
submission and submits the correct form if the form is valid.