Add support to get value of <div> elements in templates
This change is in preparation for moving the head of the form (the
shared ICS headers) into a separate file.  If that change is made, the
file will need a way to identify what form it is in for the large
header on the upper left; it can use a hidden <div> for that.