Make process_html_includes process includes sequentially
Previously, process_html_includes started a potentially large number of
asynchronous requests at once and processed them as they became
available.  Unfortunately, processing an include resulted in
invalidating the parent, so if two includes were under one parent the
second one would not be processed correctly.  This commit makes the HTML
includes be processed sequentially, and finds the element to use newly
for each element to be processed; this results in it being possible to
put multiple includes under a single parent.