Allow form field initializations to short-circuit
Previously, the form field intialization mechanism used array_for_each
to iterate over each field that might have needed an action taken on it
for a given value; this uses array_for_some instead so that functions
can short-circuit the evaluation.  This is especially useful for
"select" elements: if the value is in the "select" then the function
returns true and precludes further processing, but if it is not, the
function sets the field to "Other" and allows further processing to
happen, which results in the text field being set.