Allow Opera to be used with XHR requests
The Opera web browser will only load responseType="document" XHR
requests correctly if the MIME type of the requested document is set to
"text/xml".  This commit makes it so that if the first XHR request fails
without an error and the response type is "document" it will try again
with the MIME type overriden; this behavior can't be made default
because some other browsers will not render correctly if the document is
requested using MIME type "text/xml".