Fix handling of : in expand_template filter parsing
Prior to this commit, if a filter argument contained a colon the colon
and all subsequent filter argument text would be lost because
supplying string.split() with a limit actually splits the string fully
then just returns a number of elements less than or equal to limit.

This commit avoids that problem by splitting fully and then joining
back together the split values starting with the second one.