Move form submit buttons to persistent header
Previously form submit buttons for sending to OutPost and showing the
message text were placed at the bottom of the form where they are
often off the visible screen for large forms.

This commit places them in an absolutely positioned header bar that is
always visible no matter how the form is scrolled.  When the message
text is shown it is also done with absolute positioning so it is
visible to the user without scrolling.