Make form data message readonly and preserve focus
When message form data was shown previously it set the textarea to be
disabled to prevent the user from modifying it, which would have no
effect.  This also prevents the user from selecting the text for
copying, etc.

This commit changes to setting the textarea readOnly. It fixes up the
styling and ensures that focus moves to the message data when shown so
the user can quickly select using only the keyboard.  When the message
data is hidden again focus is switched back to where it was before.