Validate & generate output of form on user input
Previously, write_pacforms_representation() was only called when the
"Submit to Outpost" button was pressed.  This commit makes that function
be called (and, therefore, that output generated) whenever the user
inputs into the form.  Using this, the "Submit to Outpost" button is now
disabled when the form is not valid, and when it is an indication of
validity is produced: the header bar including the buttons is set to CSS
class "valid", which results in a light green background.