Fix placeholder regexp close brace handling
When the placeholders were surrounded by square brackets a single
square bracket could have ended the placeholder value and thus the
regular expression attempted to avoid matching them within the
placeholder.  When the switch was made to curly braces this was no
longer necessary but was still being used.

This commit fixes the problem and fixes one issue with IE.