Hide the form during loading and provide user feedback
Even on a local filesystem the form can take a second or so to load
and populate all the default values.  The changes to the page while
this is happening are visible.  This commit overlays the form with a
blank page initially, revealing the form once it has finished loading.
If loading takes too long it shows a spinner to keep the user
entertained.

Only one of the forms has been modified with the loading overlay while
we play with the best methods of showing and hiding the content.  The
other forms will be updated once we've finalized that.