Make loading animation fade in and out with CSS
Previously, the loading animation was enabled or disabled by JS that set
the opacity value, resulting in a sharp transition.  Now, the div is
faded in by CSS with no intervention at all from JS, and the removal of
the div is accomplished by setting the class of the element to "done"
and using a transition to fade it out.