Properly set header bar color if form is valid when loaded
Previously the header bar color was only set when something changed in
the form for the first time.  This meant that if the form was valid
when initially loaded it would not have the green color.

This commit refactors the validity check into a separate function that
can be used at load time.  Due to issues with the Javascript needing
to yield to allow the DOM to update the validity status the check is
done in a timeout.