Provide a readonly mode for viewing messages
We want to use the same form to both enter new messages and view
completed messages but it is important to keep track of which messages
are being composed and which are being viewed.

The readonly view, entered by providing a mode=readonly query
parameter makes these two state obvious.  It changes the header bar to
a blue color, hides the submission buttons because this is not a new
message, and makes all the form controls readonly so the data can not
change.  Finally it tints the form control background color to a
slightly blue gray to further emphasize the fact that the form is
readonly.