If form is valid, focus the first input element in the form
Prior to this if a valid form were loaded an attempt would be made to
focus on the undefined result of finding the first invalid field.
With this change it will focus on the first field in the form.

In addition, while testing to resolve this issue it was discovered
that, since the validity of the form won't be fully established until
the Javascript has yielded so the DOM can be updated, the
determination of the node to focus on had to be moved into the same
timeout used to change the color of the button bar.