Turn the pac-read script into a wrapper
This commit allowr pack-it-forms and PacFORMS to work at the same time
in a single installation by changing the pack-it-forms pac-read script
so that is is a wrapper for the original PacFORMS pac-read.exe.

If the script is built with cx_Freeze and installed in the PacFORMS
directory as described in the README, when called by Outpost it will
first check to see if pack-it-forms has a form with the name given by
FORMFILENAME in the message.  If it does, it will open the form using
the mechanism pack-it-forms uses.  If it does NOT, then it will call the
original PacFORMS pac-read.exe (now renamed "pac-read-pacforms.exe")
with the same arguments it was called with to give PacFORMS the chance
to handle the form.