Extract more generic selected_field_values
Prior to this commit, the function `field_value' extracted the value of
a given field in the form by using a CSS query selector in the name
attribute and then retrieving and concatenating each of the results
returned.

An upcoming commit needs the ability to select elements and retrieve
their values more flexibly using CSS selectors.  This commit extracts
the query and loop over results into a separate `selected_field_values'
function and then implements `field_value' using that function and a
join with the empty string as separator.