Additional template expansion & filters
This commit adds a new template expansion, called 'selected-fields' that
creates a list of field values matched by a CSS selector.  It also adds
filters to join the elements of a list, remove a value from a list, and
sort a list.

All of these will be used in an upcoming commit to construct message
subject automatically from some of the other values in the message.