Reorder expansions in pacforms representation
This commit is a hack to work around the fact that it is nice to be able
to auto-generate the subject of a message from the contents of that
message but the Outpost header, which itself is created by an template
expansion, needs to have the subject in it.

The commit solves this specific case by ensuring that the Outpost header
template expansion occurs after all other fields are expanded but before
any init-on-submit fields are set back to their original value.  This
ensures that the subject field is expanded before the Outpost message
header and thus the expanded value gets substituted into the Outpost
message header.