Update LA damage assessment form
Attempts to use the original form in drills showed that it was confusing
to people.  They weren't sure how to select priority locations and
requiring values for all fields in the summary resulted in a lot of
wasted typing of zeros.

In this version, there are no priority locations.  That opened up space
that is now used to allow summary stats from more than one BAT.  Some of
the summary data reported is also adjusted to be easier to use.

Finally, the message subject is now automatically generated using the
numbers of the BATs for which data is being reported, removing one more
thing for the operator to do.