Force DOM updates before continuing initialization
In the sequence of startup functions, the init_form function sets up all
the contents of form elements, etc.  These changes aren't actually
reflected in the DOM until Javascript yields, allowing the browser to
update the DOM.

Prior to this commit the only functionality in startup functions that
required the DOM contents to be updated was the validity checking.  A
setTimeout was used to delay only that processing until after the update
while allowing other initialization to proceed asynchronously.

An upcoming commit requires that the DOM contents be up to date in a
subsequent startup function.  To accommodate that and eliminate the
surprise that the contents may not be valid, this commit delays the
entire startup by 10ms before completing the init_form startup function
and calling the next one.