Add a template filter that applies expansion again
An upcoming commit will change the way that the pacforms representation
is generated to avoid issues with returning form fields to their default
contents inappropriately.

That commit will break the previous hack that ensured that the subject
field was expanded before being inserted into the Outpost header that
was required because it is sometimes desirable to automatically generate
the subject based on other message content.

This commit introduces an alternative mechanism by which the subject
field can be expanded using the templating mechanism itself, namely
adding a template expansion filter that passes the current template
value through another round of expansion.

This will be used in the future commit to retain the existing Outpost
header generation behavior.