Generate pacforms repr w/o modifying form fields
Previously, when write_pacforms_representation() expanded the templates
of init-on-submit fields, the values of the form controls were replaced
and then set back to their default values.  This had the unwanted but
unnoticed side-effect of making it impossible to edit init-on-submit
fields as write_pacforms_representation() is called every time a form
field changes.

This commit resolves the issue by expanding the values from the form
fields without replacing the existing contents so that there is no need
to restore previous values.

Because the Outpost msg header generation for init-on-submit subject
fields was relying on the expanded value being present in the subject
form field an alternative method of expanding that value was required.
The recently introduced expandtmpl template filter was used for this.