Forms in readonly mode look more like paper forms
This commit changes CSS styling in readonly mode to remove control
chrome, making the field values look like printed values on a paper
form.

Because textareas fields could contain large amounts of text and it is
difficult to style these so that they appear like simply typed text,
this commit modifies the readonly mode setup in Javascript to insert a
div element containing the text from the textarea, style it
appropriately, and hide the textarea.