Make ICS213 form position other fields required
The other value input fields for the ICS213 to/from drop-downs were not
required before which meant a form could be submitted with a To or From
field that specified Other: without actually giving a value.

This commit makes them required so that can not occur.  It works because
disabled form controls are not included in validity checking.  They are
disabled by default and only enabled when the "Other:" value of the
drop-down is selected.  If any other value is selected in the drop-down
they are cleared and disabled again.

Because different browsers fire input events differently, the event
listener for the select control value change had to be modified to
include an immediate check of the form validity whenever it was called.