Add template type for outpost envelope
Outpost adds a !OUTPOST! header line to received messages with
information about the receiver.  This commit parses those fields
out of the header and makes them available for substitution in
templates.  It will be used in upcoming commits that substitute receiver
information in some form fields to match the existing behavior of
PacFORMS.