Add expand-while-null template type
This template type takes an argument that is a list of comma separated
nested templates.  It expands them in the order they are provided until
one of the expansions has non-zero length, then returns that expansion.
The comma separating templates can be escaped with a backslash.

This capability will be used in an upcoming commit that allows the
population of fields using templates that have one template-created
value when creating a new form to be sent and a different
template-substituted value when displaying a received message.