Expand form templates after parsing message data
Prior to this commit any templates in form fields that didn't have the
class "init-on-submit" were expanded before the message data was
processed.  However, to support correct substitution of data on the
receiver's version of the form it is necessary to have access to the
form data.  This commit renames the init_empty_form() function which was
doing the template expansion to expand_templated_items() to better
reflect what it is doing and calls it after data for a message is
loaded.