Don't expand templates in values from the msg
Now that we are expanding templates in the form after loading message
data, it is possible that a field in the message might have contents
that would be interpreted as a template.  We do not want to expand such
templates as then the recipient of the message wouldn't see what the
sender actually sent.

This commit prevents the expansion by marking each element that is set
from the message data with the class 'msg-value' end then ensuring that
template expansion will not occur on any input field with that class.