Allow for form fields not set from message data
Since sender and receiver versions of forms have different data in some
fields, there are cases where we want the data for the field to come
from a template even though the message from the sender has a different
value.

Prior to this commit, if the message contained a value for the field in
the form, it would always be set to that value, which would overwrite
any template that might set it some other way.  With this commit, you
can specify that a field should not get its contents from the message by
giving the input element the class 'no-msg-init'.  It probably only
makes sense to go this if you are also specifying a default value for
the field that will give the field some other value.