Extract no-load-init from init-on-submit
Prior to this commit, if a form input element had the 'init-on-submit'
class, it would not be template expanded when the document was loaded
because this class indicates that it should instead be expanded when the
form was submitted.  However, an upcoming commit to support the
receiver-side population of form field will require some fields to be
initialized on load in the received form case and initialized on submit
in the transmitted form case.

This commit separates the behavior so that you must specify a class of
'no-load-init' to prevent expansion on page load.  If that class and the
'init-on-submit' class are both present on an input element than the
behavior will be unchanged.  If 'no-load-init' is not present then
template expansion will be done at both load and submit time.  For the
submit-time template to contain placeholders, the load-time expansion
must expand to another template.