Rename templates to escape template strings
With the change to allow nested templates the previous solution for
handling substitution of template open/close strings fails in some
situations, usually when the nested template starts or ends at another
template.

This commit renames the open brace placeholder that was used to escape
the start of a template from '{{{}}' which is hard to read and can
confuse the parsing, to '{{open-brace}}'.  It adds a corresponding
'{{close-brace}}' for situations where the closing brace needs to be
escaped.

Finally, to avoid having to specify each of these twice when two open or
close braces are desired in a row, it adds {{open-tmpl}} and
{{close-tmpl}} placeholders.