Lighten chrome around input text fields
Prior to this commit text fields had the default browser-drawn 3D style
borders.  When combined with all the horizontal and vertical rules in
the form background this made the form appear quite busy.  This commit
lightens the borders by making them a single pixel of a light gray,
attempting to retain a strong sense that these are actually input fields
while reducing the number of strong lines right next to each other.