Specify max and min widths for the form
Prior to this commit, the form would adapt in size based on the browser
width.  This was nice from a responsive design perspective, but could
lead to extremely wide or narrow forms that were completely unreadable.
This commit sets max and min sizes for the form to prevent that from
happening while allowing for a reasonable variation in form width.