Switch to CSS-based text field widths
Prior to this commit text field widths were set using the 'size'
attribute of the input element or the 'cols' element of textareas.  When
set this way, the width are fixed and can't vary as the browser window
changes, leading to alignment problems with the form.  These are even
worse when considering alternate renderings such as print.

This commit switches to specifying widths using the CSS width attribute,
which don't have these problems.  Unfortunately, to achieve the layouts
of the traditional paper forms it is necessary to specifically specify
widths on some fields, which is now done in additional CSS files
corresponding to the form or resource HTML fragment requiring the
change.

Equally unfortunately, we haven't yet uncovered a technique to place
labels on the same line as a text field that extends the rest of the
horizontal width of the form/form box.  This commit continues to punt on
that issue by specifying exact widths in em units to create
approximately the layout of the paper forms.  It would be nice to come
up with a better solution.