Supply dummy winreg module on Linux
It's handy to be able to test changes to this script on Linux but python
on Linux does not supply the winreg module.  This commit adds a dummy
implementation of the minimal functionality required by the script,
obtaining the name of a browser to open URLs.

The browser returned on Linux is hard-coded to firefox.  We tried to use
xdg-open to get the default browser, but xdg-open appears to have a bug
with file:/// URLs that contain query parameters and won't open the URL
correctly.  Since this is just a hack to make development on Linux
easier, it seems acceptable to require the availability of firefox.