pac-read.py: skip copy when file is in msgs dir
It can be convenient to use pac-read.py to open a msg that is already in
the pack-it-forms message directory in the browser.  Prior to this
commit doing this would fail because pac-read.py would attempt to copy
the file over itself, triggering an exception that the source and
destination files are the same.