pac-read.py: add --pack-it-forms-version option
This commit adds command line option called --pack-it-forms-version to
print out the installed version of pack-it-forms.  The long name for the
option and the requirement that the option is the only argument present
minimize the situations in which pac-read.py won't call through to the
wrapped PacFORMS version of pac-read.