Call wrapped pac-read.exe when < 3 cmd args
Prior to this commit, pac-read.py required at least 3 command line
arguments to be supplied or it would exit with a Traceback error.  This
is normally not a problem when Outpost is invoking pac-read.exe.
However, there are use cases where pac-read.exe is invoked manually that
use fewer arguments, most importantly in the process for manually
creating and viewing forms when the Outpost integration is not
working/available.

The solution, implemented here is to call the wrapped pac-read.exe even
when < 3 arguments are provided on the command line.

The README.md instructions for installing and updating pack-it-forms
asked the user to look for the Traceback that would be generated when
pac-read.py was called with no arguments.  Since this will no longer
happen after this commit (because pac-read.exe will be invoked with no
arguments instead) the instructions had to be updated as well.  The new
version make use of the recently added --pack-it-forms-version option,
which also has the useful property of checking the PACKITFORMS_BASE
environment variable and giving a hint if it isn't set properly.