Tweaks for better print page layout
Most importantly, eliminating the extra top margin (and ensuring
border and padding are zero) on the body that makes room for the
toolbar in the web version. Secondly, ensure that the input box chrome
is not present of buttons, comboboxes, etc.

In addition, shrinking the margins so that more information is
comfortably displayed and reducing the font size for the form data
since the printed version is higher resolution and thus it should be
easier to read.