Add array_contains() to pack-it-forms.js
This convenience function (inefficiently) determines whether an array
contains a specified value or not. It is inefficient because it is
implemented with array_for_each() and will iterate over all values
even after finding the specified value. It will be used in an
upcoming commit in which the arrays are known to be small so this
isn't currently a problem. If it becomes a problem in the future a
more efficient implementation can be provided.