Add a generic onchange handler to enable another control
Prior to this commit, the function combobox_other_manager() had been
implemented implemented to support the use of an "Other" value in a
select element by being called for the onchange event of the select
element.

An upcoming commit will add a new form that requires a similar ability
for other field types and combobox values. This abstracts
combobox_other_manager() in a function called value_based_enabler()
that takes the enabled values and target element as arguments,
re-implementing the former in terms of the latter.

An additional change to write_pacforms_representation() was required
to prevent including disabled fields in the PacFORMs representation.
This may not be strictly necessary after the bug fixes in the previous
few commits, but reduces the size of the form and shouldn't hurt.