Add a (legacy) fireEvent function
This fireEvent() function will be used in an upcoming commit to
resolve an issue where controls enabled or disabled in Javascript did
not trigger events that bubbled up to form, which meant that the
continuously updated PacFORMs representation would be temporarily out
of date with respect to the contents of the form.

The implementation uses the old deprecated Event interface to trigger
an event from Javascript. It is doing so in order to remain compatible
with IE9.