Fix Other value bug by triggering input event in Javascript
Prior to this commit there was a bug where loading a filled out form
failed to enable the "-other" input text field for the To and From ICS
Position fields when the select was set to "Other", making it
impossible to edit the string in the Other input text field. This
occurs because changing the select element's value from Javascript
doesn't trigger the onchange handler that enables the element. The
solution is to manually trigger the event.