Add button to clear the form
After submitting a form via email, the current form data remains
entered into the form, making it difficult to immediately begin
sending another form.  This behavior is desirable, as if the
communications infrastructure for email submission is not present, it
should still be possible to submit the form via Outpost.  This commit,
therefore, adds a button that makes it easy to clear the form,
intended to be used by operators after successfully submitting a form
over email.