Improve pac-read shim robustness and troubleshooting
Recently the pac-read shim has started failing on some installations
of LAARES members. Logging was not sufficient to narrow down the root
cause, necessitating a trip to examine one of the machines that was
failing. From that trip it was determined that the browser could not
be launched successfully but it wasn't clear why.

This commit makes several improvements to fix or at least allow more
effective troubleshooting of the issue.

Firstly, it replaces the way in which the browser command line to
launch the default browser is determined. Previously it would retrieve
the default browser command from the registry but just tried to
extract the path to the browser command without using the full command
template with the URL substituted as specified. With this commit I
believe it now will.

Secondly, troubleshooting was complicated by the inability to change
the command to be run to launch the browser without editing the
registry which could affect other things on the system and be
difficult to recover from if done incorrectly. This commit adds the
ability to read a browser command line template, in the same format it
is specified in the registry, from a configuration file. It also adds
the ability to enable debugging by setting a key in this file. An
example file is included in the commit.

Finally, since the information included in the log was not sufficient
to troubleshoot, this commit adds significantly more debug logging. At
startup it tries to verify the environment looks sane and suggest to
the user what to do if it does not, logging all violations. During
execution it is more explicit about the browser command line being
used and generally captures much more error information in the log.

Include the version number in the log

Add a timestamp to log messages

pac-read.py: format_browser_cmd() don't lose characters

There were two places where characters were being incorrectly omitted
from the result in the implementation. This fixes them.

Fix quoting of required modules in pac-read setup.py