pac-read.py: try harder to find browser, preferring firefox
Different versions of Windows seem to put the information about how
the default browser should be invoked. However, firefox seems to be
fairly consistent about where it writes the information about how it
would like to be invoked if it is the default under a class mapping
for FirefoxURL. Try this first in the user config, then the machine
config, and then the merged config. If none of those work, try the
same thing with the http class (the old mechanism).

If none of  these turn up a command to run, exit with an error
indicating to the user how they can modify the CONFIG file to specify
a command.