Add form version information
This commit adds version information to all forms and included HTML
fragments, which may be used in the future to ensure compatibility
between newer and older versions of pack-it-forms. The version
information of the form that produced a pack-it-forms message is
encoded in the new `# FORMVERSION` line, which records both the
version of the main form and the versions of all included HTML
fragments.