Split out outpost+pac-read integration & adjust README.md
We're planning to introduce multiple methods of using pack-it-forms
with radio applications. To keep this maintainable we're going to
maintain the integration code for different methods in separate repos.

This commit removes the files related to the only existing integration
method, which replaces the pac-read executable from PacFORMs and used
by Outpost, and updates README.md to explain where the integration
moved, etc.