Log errors during execution & show to user appropriately
Prior to this commit Javascript errors that occur during execution of
the form would often leave the form in an unusable state with no
information that might help the user resolve the problem.

This commit introduces an error log. Errors are captured from the
error event handler of the Window object and written into a div. This
approach was taken because it avoids exception handlers being
distributed all around to code to ensure that every possible error is
caught.

If an error occurs during initial page setup, the error information is
shown immediately because the form will not be usable. If an error
occurs during form data entry and/or submission the form may still be
usable to enter and transmit information so a more subtle approach is
taken. A small dark pink "Error" indicator is shown on the right hand
side of the action button toolbar. Clicking on this indicator will
toggle the visibility of error information.